Untie namespaces from protocols
Protocols no need to be based on what namespaces they look up. This fixes up
ethernet so that it doesnt need a different proto for each SNAP organizational
code.

It does mean that the namespace array is exported as a global variable so that
the lookups can be made inline. This is important because the lookups may
happen several times from within the decode hot paths.

The idea now is that decoders can be used to handle several layers as a single
logical unit (eg. tcp/ip) and to allow flexibility with performing namespace
lookups (eg. 802.1q). There is no complete decode graph within decode.c any
more.

Decoder objects will be used for extracting enpoint addresses in a uniform way
and other functions related to locating the network when there may be several
layers of tunnels and translations.

Protocol objects will be used for extracting fields, handling state tracking,
reassemblygram injection and all the other various low level protocol specific
bits and pieces.