Add basic support for protocol "decode control blocks".
Each protocol can allocate a decode control block (dcb) of a variable size up
to a declared limit. For captures where the size of decode data is unknown,
the decode stack is allocated once. A lower bound on the maximum number of
entries may be guaranteed, but for protocols with smaller decode blocks, more
layers may be decoded.

Each control block has a small header containing a pointer to the protocol
which it belongs to. It is intended that the protocol provide functions to
retrieve information from the decode control block.