Tie capture source information to packets
This is done by adding a special capture source protocol. The decode control
block simply contains a pointer to the struct _source where the packet came
from. Decided not to simply add a pointer to this in struct packet to avoid
bloating it unnecessarily, virtual re-constructed packets have no valid source.