Start work on shared memory allocator for flow trackers.
It's a two-level system. The low level part is a page allocator called memchunk
which releases pages from a fixed sized pool initiated at program startup. It
tracks a list of free pages and forgets about those which have been allocated.

The high level part is called objcache which is similar to mpool except that
it may free pages to the lower level pool, thus allowing objects of different
types to be allocated from the same global page pool. This part is not
implemented yet.