Update reassembly API.
The idea behind the new interface is that decoding of TCP application layer
protocols proceeds in three stages based on a scatter/gather I/O methodology.

Firstly, status updates are handed to application layer modules. These updates
consist of buffered data being available for processing or if one of the streams
is shutdown.

If an application module wishes to process some of the TCP buffer then it
calls back in to the reassembly module requesting for some of the buffer to be
reassembled. At this stage the associated decoder is called in to. The decoder
may, if desired, request additional data to be reassembled before the packet
containing the application protocol message is injected in to the firestorm
system via pkt_inject.

Once a packet is injected, the application layer module is given a chance to
update the flow state.

Finally, all application layer protocol handling has been moved from the
TCP state tracking module to the stream reassembly module. This is a clearer
division of funcationality and should make it easier to implement protocol
recognition heuristics in future